chore: Moved out ProcessorDistribuitor
