+++ pooler/callback_modules/module_queues_config.json
@@ -7,10 +7,6 @@
   "callback_workers": {
     "pair_total_reserves": [
       {
-          "class": "PairTotalReservesProcessorDistributor",
-          "name": "PowerLoom|UniswapPooler|Callbacks|PairTotalReservesProcessorDistributor"
-      },
-      {
         "class": "PairTotalReservesProcessor",
         "name": "PowerLoom|UniswapPooler|Callbacks|PairTotalReservesProcessor",
         "num_instances": 8
